+++ src/DevilDaggersInfo.Core.Replay/Events/Enums/BoidType.cs
@@ -1,6 +1,8 @@
 namespace DevilDaggersInfo.Core.Replay.Events.Enums;
 
+#pragma warning disable CA1008 // Enums should have zero value. Reason: Not applicable.
 public enum BoidType : byte
+#pragma warning restore CA1008
 {
 	Skull1 = 1,
 	Skull2 = 2,