+++ src/DevilDaggersInfo.Core.Replay/Events/Enums/DaggerType.cs
@@ -1,6 +1,8 @@
 namespace DevilDaggersInfo.Core.Replay.Events.Enums;
 
+#pragma warning disable CA1008 // Enums should have zero value. Reason: Not applicable.
 public enum DaggerType : byte
+#pragma warning restore CA1008
 {
 	Level1 = 1,
 	Level2 = 2,