+++ src/DevilDaggersInfo.Core.Replay/Events/Enums/PedeType.cs
@@ -1,6 +1,8 @@
 namespace DevilDaggersInfo.Core.Replay.Events.Enums;
 
+#pragma warning disable CA1008 // Enums should have zero value. Reason: Not applicable.
 public enum PedeType : byte
+#pragma warning restore CA1008
 {
 	Centipede = 1,
 	Gigapede = 2,