+++ src/DevilDaggersInfo.Core.Replay/Events/Enums/SpiderType.cs
@@ -1,6 +1,8 @@
 namespace DevilDaggersInfo.Core.Replay.Events.Enums;
 
+#pragma warning disable CA1008 // Enums should have zero value. Reason: Not applicable.
 public enum SpiderType : byte
+#pragma warning restore CA1008
 {
 	Spider1 = 1,
 	Spider2 = 2,