+++ src/DevilDaggersInfo.Core.Replay/Events/Enums/SquidType.cs
@@ -1,6 +1,8 @@
 namespace DevilDaggersInfo.Core.Replay.Events.Enums;
 
+#pragma warning disable CA1008 // Enums should have zero value. Reason: Not applicable.
 public enum SquidType : byte
+#pragma warning restore CA1008
 {
 	Squid1 = 1,
 	Squid2 = 2,