+++ src/DevilDaggersInfo.Core.Spawnset/SpawnsetSupportedGameVersion.cs
@@ -4,5 +4,8 @@ public enum SpawnsetSupportedGameVersion
 {
 	V0AndLater,
 	V2AndLater,
+#pragma warning disable CA1707
+	// ReSharper disable once InconsistentNaming
 	V3_1AndLater,
+#pragma warning restore CA1707
 }