+++ src/DevilDaggersInfo.Core.Wiki/EnemiesV2_0.cs
@@ -1,6 +1,9 @@
 namespace DevilDaggersInfo.Core.Wiki;
 
+#pragma warning disable CA1707
+// ReSharper disable once InconsistentNaming
 public static class EnemiesV2_0
+#pragma warning restore CA1707
 {
 	public static readonly Enemy Squid1 = new(GameVersion.V2_0, "Squid I", EnemyColors.Squid1, 10, 1, 2, DeathsV2_0.Purged, new(1f, 1f), 3);
 	public static readonly Enemy Squid2 = new(GameVersion.V2_0, "Squid II", EnemyColors.Squid2, 20, 2, 3, DeathsV2_0.Desecrated, new(2f, 1f), 39);