+++ src/DevilDaggersInfo.Core.Replay/Events/Data/EntityOrientationEventData.cs
@@ -11,4 +11,9 @@ public record EntityOrientationEventData(int EntityId, Int16Mat3x3 Orientation) 
 		bw.Write(EntityId);
 		bw.Write(Orientation);
 	}
+
+	public static EntityOrientationEventData CreateDefault()
+	{
+		return new(IEventData.DefaultEntityId, Int16Mat3x3.Identity);
+	}
 }