+++ src/DevilDaggersInfo.Core.Replay/Events/Data/IEventData.cs
@@ -2,5 +2,7 @@ namespace DevilDaggersInfo.Core.Replay.Events.Data;
 
 public interface IEventData
 {
+	internal static int DefaultEntityId => 1;
+
 	void Write(BinaryWriter bw);
 }