+++ src/DevilDaggersInfo.Core.Replay/Events/Data/InputsEventData.cs
@@ -28,4 +28,9 @@ public record InputsEventData(bool Left, bool Right, bool Forward, bool Backward
 		bw.Write(MouseY);
 		bw.Write((byte)0x0a);
 	}
+
+	public static InputsEventData CreateDefault()
+	{
+		return new(false, false, false, false, JumpType.None, ShootType.None, ShootType.None, 0, 0);
+	}
 }