+++ src/DevilDaggersInfo.Core.Replay/Events/Data/TransmuteEventData.cs
@@ -18,4 +18,9 @@ public record TransmuteEventData(int EntityId, Int16Vec3 A, Int16Vec3 B, Int16Ve
 		bw.Write(C);
 		bw.Write(D);
 	}
+
+	public static TransmuteEventData CreateDefault()
+	{
+		return new(IEventData.DefaultEntityId, Int16Vec3.Zero, Int16Vec3.Zero, Int16Vec3.Zero, Int16Vec3.Zero);
+	}
 }